Remove the Classic Router and simplify the Routing
